+++ sixtyfps_runtime/corelib/lib.rs
@@ -86,4 +86,5 @@ pub fn use_modules() -> usize {
         + window::ffi::sixtyfps_component_window_drop as usize
         + component::ffi::sixtyfps_component_init_items as usize
         + timers::ffi::sixtyfps_timer_start as usize
+        + graphics::color::ffi::sixtyfps_color_brighter as usize
 }